requirements.txt
torch>=2.0.0
transformers>=4.30.0
spacy>=3.5.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
